serpe: fix patterns >31 steps (progressive lengthening blank steps)
Progressive lengthening grows the pattern past ~31 steps, and the numeric
round-trip silently truncated high onsets: the appended onsets landed beyond
bit 31 and were lost/mangled, so each advance looked like it only added blank
steps.

- upi.js: bitsFromValue used JS 32-bit `>>`; the hex/octal/decimal parse used
  parseInt (Number, precision-lossy). Now BigInt throughout — unbounded length.
  Verified: 48-step pattern (onsets 0,20,39,47) round-trips exactly; short
  notation unchanged.
- C++ UPIParser: parseNumericPattern/parseDecimal used `int decimal` (32-bit);
  widened to uint64_t (binary `1ULL<<i`, hex getHexValue64, decimal
  getLargeIntValue) so the plugin round-trips patterns up to 64 steps. Verified:
  0x1000010008:40 → onsets 0,20,39 (was mangled). 134 conformance vectors still
  pass; auval clean. (>64 steps would need big-int; rare, noted.)

Webapp + plugin build.

+++ Source/Core/UPIParser.h
@@ -79,7 +79,7 @@ public:
     static std::vector<bool> parseBinary(const juce::String& binaryStr, int stepCount = 0);
     static std::vector<bool> parseArray(const juce::String& arrayStr, int stepCount = 0);
     static std::vector<bool> parseRandom(int onsets, int steps);
-    static std::vector<bool> parseDecimal(int decimal, int stepCount);
+    static std::vector<bool> parseDecimal(uint64_t decimal, int stepCount);  // 64-bit: long patterns
     static std::vector<bool> parseMorse(const juce::String& morseStr);
     static std::vector<bool> parseMorseWithDurations(const juce::String& morseStr, int shortDuration, int longDuration);
     static std::pair<std::vector<bool>, std::vector<bool>> parseMorseWithAccents(const juce::String& morseStr, char accentMode, int shortDuration = 1, int longDuration = 2);
